+++ Samples/iOS-Swift/iOS-Swift/ViewController.swift
@@ -13,6 +13,11 @@ class ViewController: UIViewController {
             let user = Sentry.User(userId: "1")
             user.email = "tony@example.com"
             scope.setUser(user)
+            
+            let path = Bundle.main.path(forResource: "Tongariro", ofType: "jpg")!
+            scope.add(Attachment(path: path, filename: "Tongariro.jpg", contentType: "image/jpeg"))
+            
+            scope.add(Attachment(data: "hello".data(using: .utf8)!, filename: "log.txt"))
         }
         // Also works
         let user = Sentry.User(userId: "1")