+++ Tests/SentryTests/Networking/TestTransport.swift
@@ -4,15 +4,15 @@ import Foundation
 public class TestTransport: NSObject, Transport {
     
     var lastSentEnvelope: SentryEnvelope?
-    var sentEvents: [Event] = []
-   
-    public func send(event: Event) {
-        sentEvents.append(event)
+    
+    var sentEvents: [Pair<Event, [Attachment]>] = []
+    public func send(event: Event, attachments: [Attachment]) {
+        sentEvents.append(Pair(event, attachments))
     }
     
-    var sentEventsWithSession: [Pair<Event, SentrySession>] = []
-    public func send(_ event: Event, with session: SentrySession) {
-        sentEventsWithSession.append(Pair(event, session))
+    var sentEventsWithSession: [Triple<Event, SentrySession, [Attachment]>] = []
+    public func send(_ event: Event, with session: SentrySession, attachments: [Attachment]) {
+        sentEventsWithSession.append(Triple(event, session, attachments))
     }
     
     var sentUserFeedback: [UserFeedback] = []